move robot types to separtate files
