+++ src/lib.rs
@@ -180,10 +180,10 @@ pub use model::PandaModel;
 pub use model::RobotModel;
 pub use network::RobotData;
 pub use robot::control_types::*;
+pub use robot::fr3::Fr3;
 pub use robot::low_pass_filter::DEFAULT_CUTOFF_FREQUENCY;
 pub use robot::low_pass_filter::MAX_CUTOFF_FREQUENCY;
+pub use robot::panda::Panda;
 pub use robot::robot_state::RobotState;
-pub use robot::Fr3;
-pub use robot::Panda;
-pub use robot::RobotWrapper;
+pub use robot::robot_wrapper::RobotWrapper;
 pub use utils::*;
