+++ src/network.rs
@@ -51,8 +51,8 @@ use crate::robot::service_types::{
     SetNeToEeRequestWithFr3Header, SetNeToEeRequestWithPandaHeader, StopMoveStatusFr3,
     StopMoveStatusPanda, FR3_VERSION, PANDA_VERSION,
 };
-use crate::robot::types::{AbstractRobotStateIntern, Fr3StateIntern, PandaStateIntern};
-use crate::{Fr3Model, PandaModel, RobotModel, RobotState};
+use crate::robot::types::AbstractRobotStateIntern;
+use crate::{RobotModel, RobotState};
 
 const CLIENT: Token = Token(1);
 
@@ -202,488 +202,6 @@ pub trait RobotData: DeviceData {
     ) -> FrankaResult<()>;
     fn handle_command_stop_move_status(status: Self::StopMoveStatus)
         -> Result<(), FrankaException>;
-}
-
-pub struct PandaData {}
-
-pub struct Fr3Data {}
-
-pub struct GripperData {}
-
-impl DeviceData for PandaData {
-    type CommandHeader = PandaCommandHeader;
-    type CommandEnum = PandaCommandEnum;
-    fn create_header(
-        command_id: &mut u32,
-        command: Self::CommandEnum,
-        size: usize,
-    ) -> Self::CommandHeader {
-        let header = PandaCommandHeader::new(command, *command_id, size as u32);
-        *command_id += 1;
-        header
-    }
-
-    fn get_library_version() -> u16 {
-        PANDA_VERSION
-    }
-}
-
-impl RobotData for PandaData {
-    type DeviceData = Self;
-    type Header = PandaCommandHeader;
-    type State = RobotState;
-    type StateIntern = PandaStateIntern;
-    type Model = PandaModel;
-    type LoadModelRequestWithHeader = LoadModelLibraryRequestWithPandaHeader;
-    type SetCollisionBehaviorRequestWithHeader = SetCollisionBehaviorRequestWithPandaHeader;
-    type SetLoadRequestWithHeader = SetLoadRequestWithPandaHeader;
-    type SetJointImpedanceRequestWithHeader = SetJointImpedanceRequestWithPandaHeader;
-    type SetCartesianImpedanceRequestWithHeader = SetCartesianImpedanceRequestWithPandaHeader;
-    type SetGuidingModeRequestWithHeader = SetGuidingModeRequestWithPandaHeader;
-    type ConnectRequestWithHeader = ConnectRequestWithPandaHeader;
-    type SetEeToKRequestWithHeader = SetEeToKRequestWithPandaHeader;
-    type SetNeToEeRequestWithHeader = SetNeToEeRequestWithPandaHeader;
-
-    type MoveRequestWithHeader = MoveRequestWithPandaHeader;
-
-    type MoveStatus = MoveStatusPanda;
-
-    type GetterSetterStatus = GetterSetterStatusPanda;
-
-    type StopMoveStatus = StopMoveStatusPanda;
-
-    type AutomaticErrorRecoveryStatus = AutomaticErrorRecoveryStatusPanda;
-
-    fn create_connect_request(
-        command_id: &mut u32,
-        udp_port: u16,
-    ) -> Self::ConnectRequestWithHeader {
-        let request = ConnectRequest::new(udp_port, PANDA_VERSION);
-        ConnectRequestWithPandaHeader {
-            header: Self::create_header(
-                command_id,
-                PandaCommandEnum::Connect,
-                size_of::<Self::ConnectRequestWithHeader>(),
-            ),
-            request,
-        }
-    }
-
-    fn create_automatic_error_recovery_request(command_id: &mut u32) -> Self::CommandHeader {
-        Self::create_header(
-            command_id,
-            PandaCommandEnum::AutomaticErrorRecovery,
-            size_of::<PandaCommandHeader>(),
-        )
-    }
-
-    fn create_stop_request(command_id: &mut u32) -> Self::CommandHeader {
-        Self::create_header(
-            command_id,
-            PandaCommandEnum::StopMove,
-            size_of::<PandaCommandHeader>(),
-        )
-    }
-
-    fn handle_command_move_status(status: Self::MoveStatus) -> Result<(), FrankaException> {
-        match status {
-            MoveStatusPanda::Success => Ok(()),
-            MoveStatusPanda::MotionStarted => {
-                //todo handle motion_generator_running == true
-                Ok(())
-            }
-            MoveStatusPanda::EmergencyAborted => Err(create_command_exception(
-                "libfranka-rs: Move command aborted: User Stop pressed!",
-            )),
-            MoveStatusPanda::ReflexAborted => Err(create_command_exception(
-                "libfranka-rs: Move command aborted: motion aborted by reflex!",
-            )),
-            MoveStatusPanda::InputErrorAborted => Err(create_command_exception(
-                "libfranka-rs: Move command aborted: invalid input provided!",
-            )),
-            MoveStatusPanda::CommandNotPossibleRejected => Err(create_command_exception(
-                "libfranka-rs: Move command rejected: command not possible in the current mode!",
-            )),
-            MoveStatusPanda::StartAtSingularPoseRejected => Err(create_command_exception(
-                "libfranka-rs: Move command rejected: cannot start at singular pose!",
-            )),
-            MoveStatusPanda::InvalidArgumentRejected => Err(create_command_exception(
-                "libfranka-rs: Move command rejected: maximum path deviation out of range!",
-            )),
-            MoveStatusPanda::Preempted => Err(create_command_exception(
-                "libfranka-rs: Move command preempted!",
-            )),
-            MoveStatusPanda::Aborted => Err(create_command_exception(
-                "libfranka-rs: Move command aborted!",
-            )),
-        }
-    }
-
-    fn create_control_exception(
-        message: String,
-        move_status: Self::MoveStatus,
-        reflex_reasons: &FrankaErrors,
-        log: Vec<Record<Self::State>>,
-    ) -> FrankaException {
-        let mut exception_string = String::from(&message);
-        if move_status == MoveStatusPanda::ReflexAborted {
-            exception_string += " ";
-            exception_string += reflex_reasons.to_string().as_str();
-            if log.len() >= 2 {
-                let lost_packets: u128 =
-                    (log.last().unwrap().state.time - log[log.len() - 2].state.time).as_millis()
-                        - 1;
-                exception_string += format!(
-                    "\ncontrol_command_success_rate: {}",
-                    log[log.len() - 2].state.control_command_success_rate
-                        * (1. - lost_packets as f64 / 100.)
-                )
-                .as_str();
-                if lost_packets > 0 {
-                    exception_string += format!(
-                        " packets lost in a row in the last sample: {}",
-                        lost_packets
-                    )
-                    .as_str();
-                }
-            }
-        }
-        FrankaException::ControlException {
-            error: exception_string,
-            log: Some(log),
-        }
-    }
-    fn create_control_exception_if_reflex_aborted(
-        message: String,
-        move_status: Self::MoveStatus,
-        reflex_reasons: &FrankaErrors,
-        log: Vec<Record<Self::State>>,
-    ) -> FrankaResult<()> {
-        // todo think about if option is a good return type
-        if move_status == MoveStatusPanda::ReflexAborted {
-            return Err(Self::create_control_exception(
-                message,
-                move_status,
-                reflex_reasons,
-                log,
-            ));
-        }
-        Ok(())
-    }
-
-    fn handle_getter_setter_status(status: Self::GetterSetterStatus) -> FrankaResult<()> {
-        match status {
-            GetterSetterStatusPanda::Success => Ok(()),
-            GetterSetterStatusPanda::CommandNotPossibleRejected => Err(create_command_exception(
-                "libfranka-rs: command rejected: command not possible in current mode",
-            )),
-            GetterSetterStatusPanda::InvalidArgumentRejected => Err(create_command_exception(
-                "libfranka-rs: command rejected: invalid argument!",
-            )),
-        }
-    }
-
-    fn handle_automatic_error_recovery_status(
-        status: Self::AutomaticErrorRecoveryStatus,
-    ) -> FrankaResult<()> {
-        match status {
-            AutomaticErrorRecoveryStatusPanda::Success => Ok(()),
-            AutomaticErrorRecoveryStatusPanda::EmergencyAborted => Err(create_command_exception(
-                "libfranka-rs: command aborted: User Stop pressed!",
-            )),
-            AutomaticErrorRecoveryStatusPanda::ReflexAborted => Err(create_command_exception(
-                "libfranka-rs: command aborted: motion aborted by reflex!",
-            )),
-            AutomaticErrorRecoveryStatusPanda::CommandNotPossibleRejected => {
-                Err(create_command_exception(
-                    "libfranka-rs: command rejected: command not possible in current mode",
-                ))
-            }
-            AutomaticErrorRecoveryStatusPanda::ManualErrorRecoveryRequiredRejected => {
-                Err(create_command_exception(
-                    "libfranka-rs: command rejected: manual error recovery required!",
-                ))
-            }
-            AutomaticErrorRecoveryStatusPanda::Aborted => {
-                Err(create_command_exception("libfranka-rs: command aborted!"))
-            }
-        }
-    }
-
-    fn handle_command_stop_move_status(
-        status: Self::StopMoveStatus,
-    ) -> Result<(), FrankaException> {
-        match status {
-            StopMoveStatusPanda::Success => Ok(()),
-            StopMoveStatusPanda::EmergencyAborted => Err(create_command_exception(
-                "libfranka-rs: Stop command aborted: User Stop pressed!",
-            )),
-            StopMoveStatusPanda::ReflexAborted => Err(create_command_exception(
-                "libfranka-rs: Stop command aborted: motion aborted by reflex!",
-            )),
-            StopMoveStatusPanda::CommandNotPossibleRejected => Err(create_command_exception(
-                "libfranka-rs: Stop command rejected: command not possible in the current mode!",
-            )),
-            StopMoveStatusPanda::Aborted => Err(create_command_exception(
-                "libfranka-rs: Stop command aborted!",
-            )),
-        }
-    }
-}
-impl DeviceData for Fr3Data {
-    type CommandHeader = Fr3CommandHeader;
-    type CommandEnum = Fr3CommandEnum;
-    fn create_header(
-        command_id: &mut u32,
-        command: Self::CommandEnum,
-        size: usize,
-    ) -> Self::CommandHeader {
-        let header = Fr3CommandHeader::new(command, *command_id, size as u32);
-        *command_id += 1;
-        header
-    }
-
-    fn get_library_version() -> u16 {
-        FR3_VERSION
-    }
-}
-impl RobotData for Fr3Data {
-    type DeviceData = Self;
-    type Header = Fr3CommandHeader;
-    type State = RobotState;
-    type StateIntern = Fr3StateIntern;
-    type Model = Fr3Model;
-    type LoadModelRequestWithHeader = LoadModelLibraryRequestWithFr3Header;
-    type SetCollisionBehaviorRequestWithHeader = SetCollisionBehaviorRequestWithFr3Header;
-    type SetLoadRequestWithHeader = SetLoadRequestWithFr3Header;
-    type SetJointImpedanceRequestWithHeader = SetJointImpedanceRequestWithFr3Header;
-    type SetCartesianImpedanceRequestWithHeader = SetCartesianImpedanceRequestWithFr3Header;
-    type SetGuidingModeRequestWithHeader = SetGuidingModeRequestWithFr3Header;
-    type ConnectRequestWithHeader = ConnectRequestWithFr3Header;
-    type SetEeToKRequestWithHeader = SetEeToKRequestWithFr3Header;
-    type SetNeToEeRequestWithHeader = SetNeToEeRequestWithFr3Header;
-    type MoveRequestWithHeader = MoveRequestWithFr3Header;
-    type MoveStatus = MoveStatusFr3;
-
-    type GetterSetterStatus = GetterSetterStatusFr3;
-
-    type StopMoveStatus = StopMoveStatusFr3;
-
-    type AutomaticErrorRecoveryStatus = AutomaticErrorRecoveryStatusFr3;
-    fn create_connect_request(
-        command_id: &mut u32,
-        udp_port: u16,
-    ) -> Self::ConnectRequestWithHeader {
-        let request = ConnectRequest {
-            version: FR3_VERSION,
-            udp_port,
-        };
-        ConnectRequestWithFr3Header {
-            header: Self::create_header(
-                command_id,
-                Fr3CommandEnum::Connect,
-                size_of::<Self::ConnectRequestWithHeader>(),
-            ),
-            request,
-        }
-    }
-
-    fn create_automatic_error_recovery_request(command_id: &mut u32) -> Self::CommandHeader {
-        Self::create_header(
-            command_id,
-            Fr3CommandEnum::AutomaticErrorRecovery,
-            size_of::<Fr3CommandHeader>(),
-        )
-    }
-
-    fn create_stop_request(command_id: &mut u32) -> Self::CommandHeader {
-        Self::create_header(
-            command_id,
-            Fr3CommandEnum::StopMove,
-            size_of::<Fr3CommandHeader>(),
-        )
-    }
-
-    fn handle_command_move_status(status: Self::MoveStatus) -> Result<(), FrankaException> {
-        match status {
-            MoveStatusFr3::Success => Ok(()),
-            MoveStatusFr3::MotionStarted => {
-            //todo handle motion_generator_running == true
-            Ok(())
-            }
-            MoveStatusFr3::EmergencyAborted => Err(create_command_exception(
-            "libfranka-rs: Move command aborted: User Stop pressed!",
-            )),
-            MoveStatusFr3::ReflexAborted => Err(create_command_exception(
-            "libfranka-rs: Move command aborted: motion aborted by reflex!",
-            )),
-            MoveStatusFr3::InputErrorAborted => Err(create_command_exception(
-            "libfranka-rs: Move command aborted: invalid input provided!",
-            )),
-            MoveStatusFr3::CommandNotPossibleRejected => Err(create_command_exception(
-            "libfranka-rs: Move command rejected: command not possible in the current mode!",
-            )),
-            MoveStatusFr3::StartAtSingularPoseRejected => Err(create_command_exception(
-            "libfranka-rs: Move command rejected: cannot start at singular pose!",
-            )),
-            MoveStatusFr3::InvalidArgumentRejected => Err(create_command_exception(
-            "libfranka-rs: Move command rejected: maximum path deviation out of range!",
-            )),
-            MoveStatusFr3::Preempted => Err(create_command_exception(
-            "libfranka-rs: Move command preempted!",
-            )),
-            MoveStatusFr3::Aborted => Err(create_command_exception(
-            "libfranka-rs: Move command aborted!",
-            )),
-            MoveStatusFr3::PreemptedDueToActivatedSafetyFunctions =>Err(create_command_exception(
-                "libfranka-rs: Move command preempted due to activated safety function! Please disable all safety functions.",
-            )),
-            MoveStatusFr3::CommandRejectedDueToActivatedSafetyFunctions =>
-                Err(create_command_exception(
-                    "libfranka-rs: Move command rejected due to activated safety function! Please disable all safety functions.",
-                ))
-
-
-        }
-    }
-
-    fn create_control_exception(
-        message: String,
-        move_status: Self::MoveStatus,
-        reflex_reasons: &FrankaErrors,
-        log: Vec<Record<Self::State>>,
-    ) -> FrankaException {
-        let mut exception_string = String::from(&message);
-        if move_status == MoveStatusFr3::ReflexAborted {
-            exception_string += " ";
-            exception_string += reflex_reasons.to_string().as_str();
-            if log.len() >= 2 {
-                let lost_packets: u128 =
-                    (log.last().unwrap().state.time - log[log.len() - 2].state.time).as_millis()
-                        - 1;
-                exception_string += format!(
-                    "\ncontrol_command_success_rate: {}",
-                    log[log.len() - 2].state.control_command_success_rate
-                        * (1. - lost_packets as f64 / 100.)
-                )
-                .as_str();
-                if lost_packets > 0 {
-                    exception_string += format!(
-                        " packets lost in a row in the last sample: {}",
-                        lost_packets
-                    )
-                    .as_str();
-                }
-            }
-        }
-        FrankaException::ControlException {
-            error: exception_string,
-            log: Some(log),
-        }
-    }
-    fn create_control_exception_if_reflex_aborted(
-        message: String,
-        move_status: Self::MoveStatus,
-        reflex_reasons: &FrankaErrors,
-        log: Vec<Record<Self::State>>,
-    ) -> FrankaResult<()> {
-        if move_status == MoveStatusFr3::ReflexAborted {
-            return Err(Self::create_control_exception(
-                message,
-                move_status,
-                reflex_reasons,
-                log,
-            ));
-        }
-        Ok(())
-    }
-
-    fn handle_getter_setter_status(status: Self::GetterSetterStatus) -> FrankaResult<()> {
-        match status {
-                GetterSetterStatusFr3::Success => Ok(()),
-                GetterSetterStatusFr3::CommandNotPossibleRejected => Err(create_command_exception(
-                    "libfranka-rs: command rejected: command not possible in current mode",
-                )),
-                GetterSetterStatusFr3::InvalidArgumentRejected => Err(create_command_exception(
-                    "libfranka-rs: command rejected: invalid argument!",
-                )),
-                GetterSetterStatusFr3::CommandRejectedDueToActivatedSafetyFunctions => Err(create_command_exception(
-                    "libfranka-rs: command rejected due to activated safety function! Please disable all safety functions.",
-                )),
-            }
-    }
-
-    fn handle_automatic_error_recovery_status(
-        status: Self::AutomaticErrorRecoveryStatus,
-    ) -> FrankaResult<()> {
-        match &status {
-            AutomaticErrorRecoveryStatusFr3::Success => Ok(()),
-            AutomaticErrorRecoveryStatusFr3::EmergencyAborted => Err(create_command_exception(
-                "libfranka-rs: command aborted: User Stop pressed!",
-            )),
-            AutomaticErrorRecoveryStatusFr3::ReflexAborted => Err(create_command_exception(
-                "libfranka-rs: command aborted: motion aborted by reflex!",
-            )),
-            AutomaticErrorRecoveryStatusFr3::CommandNotPossibleRejected => {
-                Err(create_command_exception(
-                    "libfranka-rs: command rejected: command not possible in current mode",
-                ))
-            }
-            AutomaticErrorRecoveryStatusFr3::ManualErrorRecoveryRequiredRejected => {
-                Err(create_command_exception(
-                    "libfranka-rs: command rejected: manual error recovery required!",
-                ))
-            }
-            AutomaticErrorRecoveryStatusFr3::Aborted => {
-                Err(create_command_exception("libfranka-rs: command aborted!"))
-            }
-            AutomaticErrorRecoveryStatusFr3::CommandRejectedDueToActivatedSafetyFunctions => Err(create_command_exception(
-                "libfranka-rs: command rejected due to activated safety function! Please disable all safety functions.",
-            )),
-        }
-    }
-
-    fn handle_command_stop_move_status(
-        status: Self::StopMoveStatus,
-    ) -> Result<(), FrankaException> {
-        match status {
-            StopMoveStatusFr3::Success => Ok(()),
-            StopMoveStatusFr3::EmergencyAborted => Err(create_command_exception(
-                "libfranka-rs: Stop command aborted: User Stop pressed!",
-            )),
-            StopMoveStatusFr3::ReflexAborted => Err(create_command_exception(
-                "libfranka-rs: Stop command aborted: motion aborted by reflex!",
-            )),
-            StopMoveStatusFr3::CommandNotPossibleRejected => Err(create_command_exception(
-                "libfranka-rs: Stop command rejected: command not possible in the current mode!",
-            )),
-            StopMoveStatusFr3::Aborted => Err(create_command_exception(
-                "libfranka-rs: Stop command aborted!",
-            )),
-            StopMoveStatusFr3::CommandRejectedDueToActivatedSafetyFunctions => Err(create_command_exception(
-                "libfranka-rs: Move command rejected due to activated safety function! Please disable all safety functions.",
-            ))
-        }
-    }
-}
-
-impl DeviceData for GripperData {
-    type CommandHeader = GripperCommandHeader;
-    type CommandEnum = GripperCommandEnum;
-
-    fn create_header(
-        command_id: &mut u32,
-        command: Self::CommandEnum,
-        size: usize,
-    ) -> Self::CommandHeader {
-        let header = GripperCommandHeader::new(command, *command_id, size as u32);
-        *command_id += 1;
-        header
-    }
-
-    fn get_library_version() -> u16 {
-        GRIPPER_VERSION
-    }
 }
 
 pub trait MessageCommand {
